Update README.md
SQL (Structured Query Language) is a language for interacting with data stored in relational database.
Relational database is a collection of tables. Each row (record) of a table contains a single entity. Each column (field) of a table contains a single attribute.
Query is a request for data from from a database table.

Sidenotes:
1. COMMANDS are not case-sensitive
2. Include a semicolon at the end of query